Internals

Contents

Internals

Contents

Index

Internals

Index

MEstimation.MEstimation_results

Base.show

Internals

MEstimation.MEstimation_results — Type

MEstimation_results(results::Union{NLsolve.SolverResults, Optim.MultivariateOptimizationResults, Optim.UnivariateOptimizationResults},
                    theta::Vector,
                    data::Any,
                    template::Union{objective_function_template, estimating_function_template},
                    regularizer::Function,
                    br::Bool,
                    has_objective::Bool,
                    has_regularizer::Bool,
                    br_method::String)

Composite type for the output of fit for an objective_function_template or an estimating_function_template.

Arguments

results: a Union{NLsolve.SolverResults, Optim.MultivariateOptimizationResults, Optim.UnivariateOptimizationResults} object holding the optimization results as returned from Optim.optimize or NLsolve.nlsolve.

theta: a Vector{Float64} with the M-estimates or their reduced-bias version.

data: an object of composite type with all the data required to compute the objective function or the estimating functions; see objective_functionand [estimating_function`](@ref).

template: an objective_function_template or estimating_function_template object.

regularizer: a function of theta and data returning either a Vector{Float64} of dimension equal to the number of the estimating functions (for estimating_function_template) or a Float64 (for objective_function_template). See fit for details.

br: a Bool; if false then results are from the computation of M-estimates, otherwise from the computation of the RBM-estimates.

has_objective: a Bool; if true then template is an objective_function_template.

has_regularizer: a Bool; if true then a regularizer function has been used during optimization; see fit.

br_method: either "implicittrace" (default) or "explicittrace"; see fit.

Base.show — Function

show(io::IO, 
     results::MEstimation_results; 
     digits::Int64 = 4)

show method for MEstimation_results objects.

Arguments

io: an IO object; see show for details.

results: a MEstimation_results object.

Keyword arguments

digits: an Int64 indicating the number of digits to display for the various summaries. Default is 4.

Details

If MEstimation_results.has_objective == true, then the result of aic(results) and tic(results) are also printed.